gradient

+++ Backend/Image/Image.h
@@ -6,9 +6,9 @@ class Image {
  public:
    struct Pixel
    {
-      uint8_t r = 0;
-      uint8_t g = 0;
-      uint8_t b = 0;
+      double r = 0;
+      double g = 0;
+      double b = 0;
    };
 
  private:
